Add a safety checker for actions defined in interactions.
Every action that can be called by an interaction should be present in
on of the action files. Either, the game could crash since trying to
call an inexistent action. It is not complete, since it does not check
that targets are compatible, however it was the least to check.